divide component
