+++ core/ui/src/main/java/com/segnities007/ui/indicator/CircularProgressWithPercentage.kt
@@ -1,19 +1,14 @@
 package com.segnities007.ui.indicator
 
-import androidx.compose.foundation.background
-import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.BoxWithConstraints
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.size
-import androidx.compose.foundation.shape.RoundedCornerShape
 import androidx.compose.material3.CircularProgressIndicator
 import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.draw.clip
-import androidx.compose.ui.graphics.Brush
 import androidx.compose.ui.graphics.StrokeCap
 import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.unit.dp
@@ -28,10 +23,11 @@ fun CircularProgressWithPercentage(
         contentAlignment = Alignment.Center,
         modifier = modifier,
     ) {
-        val size = maxWidth
-        val stroke = size * 0.08f
-        val fontSize = size * 0.24f
+        val size = maxWidth // 親のサイズ
+        val stroke = size * 0.12f // Material3 Expressive: サイズの12%をストローク幅に
+        val fontSize = size * 0.24f // サイズの24%をフォントサイズに
 
+        // プログレスインジケータ - Material3 Expressive
         CircularProgressIndicator(
             progress = { progress },
             strokeWidth = stroke,
@@ -39,6 +35,7 @@ fun CircularProgressWithPercentage(
             strokeCap = StrokeCap.Round,
         )
 
+        // 中央のパーセント表示
         Text(
             text = "${(progress * 100).toInt()}%",
             fontSize = fontSize.value.sp,
